Using logic from cf-xarray to streamline vars
Previously variables have had to be input for each reader in particular. The changes in this branch, and started in this commit, avoid this problem by using some logic from cf-xarray that using a user-defined dictionary of ways to identify a variable. Going forward, a user will be able to input the specific reader variable names if they prefer, but won't have to.