[OSRA-88] Patch vuln: using ast on untrusted sources
ast by itself on _trusted_ sources is ok according to the docs. In this
case however, ast.literal_eval is being called on an untrusted document
which can be exploited. I tested locally and found an exploit with
minimal effort, so it's not safe. This commit changes the function away
from reading python literals to reading JSON which even from untrusted
sources is more secure. Since the object expected from the source is
like a Python dictionary, transitioning to JSON from a python source
file should be relatively straightforward.